Move to TypeScript modules for `.py` files
