+++ web/src/global.d.ts
@@ -1,0 +1,4 @@
+declare module '*.py' {
+  const content: any;
+  export default content;
+}
